requirements.txt
discord.py[voice]
python-dotenv
aiohttp
gTTS
